pyproject.toml
[project]
name = "web-docs-editor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "llama-cpp-python",
    "uvicorn",
]
